requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.6
django-htmx==1.17.0
django-jsonform==2.19.1
django-phonenumber-field==7.2.0
django-user-sessions==2.0.0
frozenlist==1.4.0
geoip2==4.7.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
maxminddb==2.4.0
multidict==6.0.4
phonenumbers==8.13.23
Pillow==10.1.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
twilio==8.10.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
yarl==1.9.2
zope.interface==6.1
